terraform plugin protocol v6
This commit introduces the new plugin protocol version 6. This version
supports NestedTypes inside Attributes, which allows provider developers
to define object-type Attributes using a NestedBlock syntax.

The new plugin6 package handles the conversion between the protocol
(tfplugin6) and configschema. There will be some more minor changes to
the protocol in future commits; the goal of this commit is to get the bulk
of the changes in place before getting started on the changes that will
end up touching many files.