command/format: a tiny start on tests and TODOs for attrs with nested types
This commit is a small bit of setup for upcoming work on the format
printer for attributes with NestedTypes. I added an attribute with
NestedType + NestingMode NestingList to the
TestResourceChange_nestedList suite of tests to confirm that the output
looks good enough for now, and added some comments about what still
needs to be done and what tests _should_ fail once the diff printing is
improved.

This is admittedly not a very high-value commit, but I'd rather get the
tests committed now than risk losing them; these are especially tedious
to hand-craft.